Refactor: AdminUserController 리펙토링 완료(#82)

+++ src/main/java/com/sparta/ordermanagement/bootstrap/admin/controller/AdminUserController.java
@@ -1,22 +1,20 @@
 package com.sparta.ordermanagement.bootstrap.admin.controller;
 
 import com.sparta.ordermanagement.application.admin.AdminUserService;
-import com.sparta.ordermanagement.application.exception.InvalidValueException;
 import com.sparta.ordermanagement.bootstrap.admin.dto.UserAdminSignupRequest;
 import com.sparta.ordermanagement.bootstrap.admin.dto.UserGetResponse;
 import com.sparta.ordermanagement.bootstrap.auth.UserDetailsImpl;
-import com.sparta.ordermanagement.bootstrap.rest.exception.exceptions.InvalidAuthorizationException;
+import com.sparta.ordermanagement.bootstrap.rest.exception.exceptions.ForbiddenActionException;
+import com.sparta.ordermanagement.bootstrap.rest.exception.exceptions.RequestValidationException;
 import com.sparta.ordermanagement.framework.persistence.entity.user.Role;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.security.sasl.AuthenticationException;
 import java.util.List;
 import java.util.Map;
 
@@ -28,33 +26,34 @@ public class AdminUserController {
 
     private final AdminUserService adminUserService;
 
+    private static void validateUserRoleForAccess(UserDetailsImpl userDetails) {
+        if(!(userDetails.getRole() == Role.MANAGER || userDetails.getRole() == Role.MASTER)) {
+            throw new ForbiddenActionException("API 접근 권한이 없습니다.");
+        }
+    }
+
     @PostMapping("")
     @ResponseStatus(HttpStatus.CREATED)
-    public String createAdmin(
-            @RequestBody UserAdminSignupRequest userAdminSignupRequest, BindingResult bindingResult,
+    public Map.Entry createAdmin(
+            @RequestBody @Valid UserAdminSignupRequest userAdminSignupRequest, BindingResult bindingResult,
             @AuthenticationPrincipal UserDetailsImpl userDetails
     ) {
 
-        log.info("[AdminUserController] /admin/v1/users call");
+        log.info("[AdminUserController]-[createAdmin] API call");
 
-        if(userDetails.getAuthorities() == null || userDetails.getAuthorities().isEmpty()) {
-            throw new InvalidAuthorizationException("Authorities cannot be null or empty");
-        }
-
-        // 유효성 검사 실패 시 오류 처리
+        // Validation
         if (bindingResult.hasErrors()) {
-
-            StringBuilder errorMessages = new StringBuilder();
-            bindingResult.getAllErrors().forEach(error -> {
-                errorMessages.append(error.getDefaultMessage()).append("\n");
-            });
-
-            log.error("[AdminUserController] /admin/v1/users error: {}", errorMessages.toString());
-
-            throw new InvalidValueException(errorMessages.toString());
+            String bindingErrorMessage = bindingResult.getAllErrors().toString();
+            throw new RequestValidationException(bindingErrorMessage);
+        }
+        if(!(userDetails.getRole() == Role.MANAGER || userDetails.getRole() == Role.MASTER || userDetails.getRole() == Role.OWNER)) {
+            //이후 OWNER 삭제하기
+            throw new ForbiddenActionException("API 접근 권한이 없습니다.");
         }
 
-        return adminUserService.signup(userAdminSignupRequest.toDomain());
+        String signupUserStringId = adminUserService.signup(userAdminSignupRequest.toDomain());
+
+        return Map.entry("userStringId", signupUserStringId);
     }
 
 
@@ -62,7 +61,10 @@ public class AdminUserController {
     @ResponseStatus(HttpStatus.OK)
     public List<UserGetResponse> getUsers(@AuthenticationPrincipal UserDetailsImpl userDetails){
 
-        log.info("[AdminUserController]-[getUsers] /admin/v1/users call");
+        log.info("[AdminUserController]-[getUsers] API call");
+
+        // Validation
+        validateUserRoleForAccess(userDetails);
 
         return adminUserService.getUsers().stream()
                 .map(UserGetResponse::from)
@@ -76,10 +78,14 @@ public class AdminUserController {
             @PathVariable("role") Role role,
             @AuthenticationPrincipal UserDetailsImpl userDetails){
 
-        log.info("[AdminUserController]-[updateUser] /admin/v1/users/{user_string_id} call");
-        Integer updatedUserId = adminUserService.updateUser(userStringId, role);
+        log.info("[AdminUserController]-[updateUser] API call");
 
-        return Map.entry("userId", updatedUserId);
+        // Validation
+        validateUserRoleForAccess(userDetails);
+
+        adminUserService.updateUser(userStringId, role);
+
+        return Map.entry("userStringId", userStringId);
     }
 
     @DeleteMapping("/{user_string_id}")
@@ -88,11 +94,14 @@ public class AdminUserController {
             @PathVariable("user_string_id") String userStringId,
             @AuthenticationPrincipal UserDetailsImpl userDetails){
 
-        log.info("[AdminUserController]-[deleteUser] /admin/v1/users/{} call", userStringId);
+        log.info("[AdminUserController]-[deleteUser] API call");
+
+        // Validation
+        validateUserRoleForAccess(userDetails);
 
         adminUserService.deleteUser(userStringId);
 
-        return Map.entry("userId", userStringId);
+        return Map.entry("userStringId", userStringId);
     }
 
 }
